feat : AnswerController 에서 BindingResult 를 이용한 에러처리 구현

+++ src/main/java/com/peer/missionpeerflow/controller/AnswerController.java
@@ -2,6 +2,7 @@ package com.peer.missionpeerflow.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -24,19 +25,28 @@ public class AnswerController {
     private final AnswerService answerService;
 
     @PostMapping("")
-    public ResponseEntity create(@Valid @RequestBody AnswerCreateDTO answerCreateDTO) {
+    public ResponseEntity create(@Valid @RequestBody AnswerCreateDTO answerCreateDTO, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(SuccessMessage.of("Error occurred while Request Body Validation"));
+        }
         answerService.create(answerCreateDTO);
         return ResponseEntity.status(HttpStatus.CREATED).body(SuccessMessage.of("Answer created successfully"));
     }
 
     @PutMapping("/{answerId}")
-    public ResponseEntity modify(@Valid @RequestBody AnswerModifyDTO answerModifyDTO, @PathVariable(name = "answerId") Long id) {
+    public ResponseEntity modify(@Valid @RequestBody AnswerModifyDTO answerModifyDTO, @PathVariable(name = "answerId") Long id, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(SuccessMessage.of("Error occurred while Request Body Validation"));
+        }
         answerService.modify(answerModifyDTO, id);
         return ResponseEntity.status(HttpStatus.CREATED).body(SuccessMessage.of("Answer modified successfully"));
     }
 
     @PostMapping("/{answerId}")
-    public ResponseEntity delete(@Valid @RequestBody AnswerDeleteDTO answerDeleteDTO, @PathVariable(name = "answerId") Long id) {
+    public ResponseEntity delete(@Valid @RequestBody AnswerDeleteDTO answerDeleteDTO, @PathVariable(name = "answerId") Long id, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(SuccessMessage.of("Error occurred while Request Body Validation"));
+        }
         answerService.delete(answerDeleteDTO, id);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).body(SuccessMessage.of("Answer deleted successfully"));
     }
